fix: move install-theme.js to lib directory
