+++ lib/install-theme.js
